requirements.txt
# Django & Core
Django>=5.0.1,<5.1
django-cors-headers>=4.3.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9

# REST API
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5

# WebSockets & ASGI
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
asgiref>=3.7.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# File Storage (local + Cloudflare R2 / S3)
django-storages>=3.14.0
boto3>=1.34.0

# Security & Encryption
cryptography>=42.0.0

# Validation & Utilities
phonenumbers>=8.13.0

# AI Services
openai>=1.12.0
